% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_call_functions.r
\name{binnedSum}
\alias{binnedSum}
\title{binnedSum: this function computes the coverage over a binned chromosome,
starting from a per base computed coverage
credits: http://crazyhottommy.blogspot.com/2016/02/compute-averagessums-on-granges-or.html}
\usage{
binnedSum(bins, numvar, mcolname)
}
\arguments{
\item{bins}{a GRanges object representing a chromosome binned}

\item{numvar}{an RleList representing the per base coverage over the chr}

\item{mcolname}{the name of column where the sum have to be stored}
}
\value{
the bins GRanges with the mcolname attached
}
\description{
binnedSum: this function computes the coverage over a binned chromosome,
starting from a per base computed coverage
credits: http://crazyhottommy.blogspot.com/2016/02/compute-averagessums-on-granges-or.html
}
\examples{
## dividing one chromosome in bins of 50 bp each
seqinfo <- GenomeInfoDb::Seqinfo("mm9")
bins <- GenomicRanges::tileGenome(seqlengths=seqinfo@chr1,
                                             tilewidth=50,
                                             cut.last.tile.in.chrom=TRUE)
gr <- GRanges(seqnames = Rle("Chr1", 100),
ranges =IRanges(start = seq(from=10, to=1000, by=10),
end=seq(from=20, to=1010, by = 10)))
## computing coverage per single base on granges
cov <- GenomicRanges::coverage(x=gr)

binnedCovGR <- binnedSum(bins, cov, "binned_cov")

}
