% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_call_functions.r
\name{binnedSumOnly}
\alias{binnedSumOnly}
\title{binnedSumOnly}
\usage{
binnedSumOnly(bins, numvar, mcolname)
}
\arguments{
\item{bins}{a GRanges object representing a chromosome binned}

\item{numvar}{an RleList representing the per base coverage over the chr}

\item{mcolname}{the name of column where the sum have to be stored}
}
\value{
an Rle within the per bin computed coverage
}
\description{
it's useful just to coerce the bin coverage to an Rle object
}
