% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_region_functions.R
\name{catRegions}
\alias{catRegions}
\title{concatenate region files from various chromosomes into a genome
wide region file.}
\usage{
catRegions(path = "FinalRegions", type = "RData", keep_files = TRUE)
}
\value{
c.
}
\description{
concatenate region files from various chromosomes into a genome
wide region file.
}
\keyword{internal}
