% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakCallFunctions.R
\name{computeCoverageMovingWindowOnChr}
\alias{computeCoverageMovingWindowOnChr}
\title{computeCoverageMovingWindowOnChr}
\usage{
computeCoverageMovingWindowOnChr(chrBedGRanges, minWinWidth = 50,
  maxWinWidth = 1000, binWidth = 50, verbose = TRUE)
}
\arguments{
\item{chrBedGRanges}{a GRanges to compute the coverage}

\item{minWinWidth}{the minimum width of the window to use for the coverage}

\item{maxWinWidth}{the maximum width of the window to use for the coverage}

\item{binWidth}{the dimension of the bin in base number}
}
\value{
RleList where each element is a window within the Rle of its coverage
}
\description{
computes the coverage on a chromosomewith a
set of moving windows of dimensions minWinWidth:maxWinWidth
}
\keyword{internal}
