% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{constructBedRanges}
\alias{constructBedRanges}
\title{constructBedRanges}
\usage{
constructBedRanges(filename, filetype = c("bam", "bed"), genomeName = NULL,
  onlyStdChrs = FALSE)
}
\arguments{
\item{filename}{the complete file path of a bam?bed file}

\item{filetype}{the file type bam/bed}

\item{genomeName}{the name of the genome used to map the reads (i.e. mm9)
N.B. if NOT NULL the GRanges Seqinfo will be forced to
genomeName Seqinfo}

\item{onlySrdChrs}{a flag to keep only standard chromosomes from the input
files.}
}
\value{
a GRanges object
}
\description{
Constructs a GRanges object from a bam/bed file in a consistent way
}
\keyword{internal}
