% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEScan2.R
\docType{package}
\name{DEScan2}
\alias{DEScan2}
\alias{DEScan2-package}
\title{DEScan2}
\description{
integrated peak and differential caller,
specifically designed for broad epigenomic signals.
}
\author{
some authors
}
