% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{divideEachSampleByChromosomes}
\alias{divideEachSampleByChromosomes}
\title{divideEachSampleByChromosomes}
\usage{
divideEachSampleByChromosomes(samplesGRangesList)
}
\arguments{
\item{samplesGRangesList}{a GRangesList of samples}
}
\value{
list of samples where each element is a list of chromosomes and each
         of these elements is a granges
}
\description{
taken in input a grangeslist of samples, generate a list of samples
             where each element has a GRangesList each element of the
             GRangesList represents a single chromosome
}
\examples{
TBW
}
