% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_region_functions.R
\name{findOverlapsOverSamples}
\alias{findOverlapsOverSamples}
\title{findOverlapsOverSamples}
\usage{
findOverlapsOverSamples(samplePeaksGRangelist, extendRegions = 200,
  minOverlap = 0L, maxGap = -1L)
}
\arguments{
\item{samplePeaksGRangelist}{given a granges list of samples finds
the overlapping regions between them}

\item{extendRegions}{the number of bases to extend each region at its start
and end}

\item{minOverlap}{the minimum overlap each peak needs to have
(see ChipPeakAnno::findOverlapsOfPeaks)}

\item{maxGap}{the maximum gap admissible between the peaks
(see ChipPeakAnno::findOverlapsOfPeaks)}
}
\value{
a GRanges of peaks overlapped and unique between samples
}
\description{
given in input a GRangeList where each element is a sample
             computes the coverage extending a both direction window of
             prefixed length.
}
\examples{
TBW
}
