% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fromSamplesToChromosomesGRangesList}
\alias{fromSamplesToChromosomesGRangesList}
\title{fromSamplesToChromosomesGRangesList}
\usage{
fromSamplesToChromosomesGRangesList(samplesGRangesList)
}
\arguments{
\item{samplesGRangesList}{a GRangesList of samples.
Tipically generaed by findPeaks}
}
\value{
A GRangesList of chromosomes where each element is a GRanges list
        of samples
}
\description{
converts a GRangesList orgnized per samples to a GRangesList
             organized per Chromosomes where each element
             is a GRangesList of samples
}
\examples{
TBW
}
