% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generateDFofSamplesPerChromosomes}
\alias{generateDFofSamplesPerChromosomes}
\title{generateDFofSamplesPerChromosomes}
\usage{
generateDFofSamplesPerChromosomes(samplesChrGRList)
}
\arguments{
\item{samplesChrGRList}{a GRangesList of samples each divided by chromosome}
}
\value{
a dataframe  where each row is a sample (1st col) and
        a string with its chromosomes separated by ";" (2nd col)
}
\description{
generates a dataframe where each row is a sample (1st col) and
             a string with its chromosomes separated by ";" (2nd col)
             (useful to fromSamplesToChromosomesGRangesList function)
}
