% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_region_functions.R
\name{giveUniqueNamesToPeaksOverSamples}
\alias{giveUniqueNamesToPeaksOverSamples}
\title{giveUniqueNamesToPeaksOverSamples}
\usage{
giveUniqueNamesToPeaksOverSamples(samplePeaksGRangelist)
}
\arguments{
\item{samplePeaksGRangelist}{}
}
\value{
a GRangesList of samples within renamed peaks for each element
}
\description{
given a GRangesList of samples assigns unique names to the peaks
             of each sample
}
\keyword{internal}
