% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalRegionFunctions.R
\name{initMergedPeaksNames}
\alias{initMergedPeaksNames}
\title{initMergedPeaksNames}
\usage{
initMergedPeaksNames(mergedGRanges)
}
\arguments{
\item{mergedGRanges}{A GRanges object.
(Tipically Generated in findOverlapsOverSamples function )}
}
\value{
a granges of renamed peaks.
}
\description{
given a GRanges of merged peaks assigns them new names.
}
\keyword{internal}
