% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{keepRelevantChrs}
\alias{keepRelevantChrs}
\title{keepRelevantChrs}
\usage{
keepRelevantChrs(GRangesList, chr)
}
\arguments{
\item{GRangesList}{where each element is a chromosome,
tipically created with cutGRangesPerChromosome}

\item{chr}{a character vector of chromosomes names of the form "chr#"}
}
\value{
the input GRangesList with only the relevat chromosomes
}
\description{
subselect a list of GRanges created with
cutGRangesPerChromosome returning only the relevant chromosomes GRanges
}
\examples{
TBW
}
