% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_call_functions.R
\name{make_grid}
\alias{make_grid}
\title{make grid for enrichment scan.}
\usage{
make_grid(bed, fraglen, rlen, min_bin = 10, min_rds = 0, grid_st = NA,
  grid_ed = NA)
}
\arguments{
\item{fraglen}{Integer indicating the average DNA fragment length in bp}

\item{rlen}{Integer indicating the read length in bp}

\item{path}{Path to alignement files.}

\item{chr}{Integer specifying which chromosme to limit analysis to.}

\item{min_win}{Integer indicating the minimum window size to be used in constructing grid}

\item{max_win}{Integer indicating the maximum window size to be used in constructing grid}

\item{blocksize}{Integer indicating how large of a chunk of the chromosome will be analyzed at a time}
}
\value{
grid.
}
\description{
make grid for enrichment scan.
}
\keyword{internal}

