% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_call_functions.r
\name{makeGrid}
\alias{makeGrid}
\title{make grid for enrichment scan.}
\usage{
makeGrid(bed, fragmentLength, readLength, minBin = 10, minRds = 0,
  gridStart = NA, gridEnd = NA, verbose = FALSE)
}
\arguments{
\item{gridStart}{Genomic Position to start the grid}

\item{path}{Path to alignement files.}

\item{fraglen}{Integer indicating the average DNA fragment length in bp}

\item{rlen}{Integer indicating the read length in bp}

\item{min_bin}{Integer indicating the minimum window size to be used in
constructing grid}

\item{min_rds}{This argument is currently not used.}

\item{grid_ed}{Genomic Position to end the grid}
}
\value{
grid.
}
\description{
make grid for enrichment scan.
}
\keyword{internal}
