% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_region_functions.R
\name{merge_overlapping_intervals}
\alias{merge_overlapping_intervals}
\title{merge overlapping peaks across replicates.}
\usage{
merge_overlapping_intervals(s)
}
\value{
s2.
}
\description{
merge overlapping peaks across replicates.
}
\keyword{internal}

