% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpparma_get_disjoint_max_win}
\alias{rcpparma_get_disjoint_max_win}
\title{rcpparma_get_disjoint_max_win
it computes the disjoint max_win matrix}
\usage{
rcpparma_get_disjoint_max_win(z0, sigwin = 20L, zthresh = -9999999,
  nmax = 9999999L, verbose = TRUE)
}
\arguments{
\item{z0}{a matrix}

\item{sigwin}{sigwin}

\item{zthresh}{zthresh}

\item{nmax}{nmax}

\item{verbose}{verbose}
}
\description{
rcpparma_get_disjoint_max_win
it computes the disjoint max_win matrix
}
\keyword{internal}
