% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_peaks.R
\name{read_bam}
\alias{read_bam}
\title{read a bam file into a bed like format.}
\usage{
read_bam(file, chr)
}
\arguments{
\item{file}{Character indicating path to bam file.}

\item{chr}{Integer indicating which chromosome to read in.}
}
\value{
bed.
}
\description{
read a bam file into a bed like format.
}
\keyword{internal}
