% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{readBamAsBed}
\alias{readBamAsBed}
\title{readBamAsBed}
\usage{
readBamAsBed(file)
}
\arguments{
\item{file}{Character indicating path to bam file.}
}
\value{
GRanges object.
}
\description{
read a bam file into a bed like format.
              forcing UCSC format for chromosomes names.
}
\keyword{internal}
