% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{readBedFile}
\alias{readBedFile}
\title{readBedFile}
\usage{
readBedFile(filename)
}
\arguments{
\item{filename}{}
}
\value{
GRanges object
}
\description{
read a bed file into a GenomicRanges like format.
             forcing UCSC format for chromosomes names.
}
\keyword{internal}
