% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{RleListToRleMatrix}
\alias{RleListToRleMatrix}
\title{RleListToRleMatrix: a wrapper to create a RleMatrix from a RleList object}
\usage{
RleListToRleMatrix(RleList, dimnames = NULL)
}
\arguments{
\item{RleList}{an RleList object}

\item{dimnames}{the names for dimensions of RleMatrix (see DelayedArray pkg)}
}
\value{
a RleMatrix from DelayedArray package
}
\description{
RleListToRleMatrix: a wrapper to create a RleMatrix from a RleList object
}
