% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_peaks.R
\name{window_coverage_old}
\alias{window_coverage_old}
\title{compute window coverage.}
\usage{
window_coverage_old(Fr, Rr, grid, fraglen = 200, rlen = 100, min_win = 1,
  max_win = 100, verbose = TRUE)
}
\arguments{
\item{Fr}{forward reads.}

\item{Rr}{reverse reads.}
}
\value{
cmat.
}
\description{
compute window coverage.
}
\keyword{internal}
