% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_call_functions.r
\name{computeLambdaOnChr}
\alias{computeLambdaOnChr}
\title{computeLambdaOnChr: computes the lambdas on a chromosome for the
winVector windows and other two windows (min/maxCompWinWidth) to compare with}
\usage{
computeLambdaOnChr(chrGRanges, winVector = c(1:20), minChrRleWComp,
  minCompWinWidth = 5000, maxChrRleWComp, maxCompWinWidth = 10000)
}
\arguments{
\item{chrGRanges}{the GRanges representing the reads of the chromosome}

\item{winVector}{the of width of the windows used to compute the coverage}

\item{minChrRleWComp}{and Rle object within coverage of window of width
minCompWinWidth}

\item{minCompWinWidth}{the width of the window used for the coverage of
minChrRleWComp}

\item{maxChrRleWComp}{and Rle object within coverage of window of width
minCompWinWidth}

\item{maxCompWinWidth}{the width of the window used for the coverage of
maxChrRleWComp}
}
\value{
an RleList where each element is a window of winVector, within an Rle
        representing the lambda computed for that window
}
\description{
computeLambdaOnChr: computes the lambdas on a chromosome for the
winVector windows and other two windows (min/maxCompWinWidth) to compare with
}
