% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrixFunctions.R
\name{countFinalRegions}
\alias{countFinalRegions}
\title{countFinalRegions}
\usage{
countFinalRegions(regionsGRanges, readsFilePath = NULL, fileType = c("bam",
  "bed"), minCarriers = 2, genomeName = NULL, onlyStdChrs = FALSE,
  ignStrand = TRUE, saveFlag = FALSE, savePath = "finalRegions")
}
\arguments{
\item{regionsGRanges}{a GRanges objects representing the peaks to compute
the coverage, with a "k-carriers" mcols.
(tipically generated by finalRegions function).}

\item{readsFilePath}{the filepath of bam or bed files necessary to compute
the coverage.}

\item{fileType}{the fileType of the input files.}

\item{minCarriers}{minimum number of carriers (samples).}

\item{genomeName}{code name of the genome of reads files (optional).}

\item{onlyStdChrs}{a flag indicating if to keep only the standard chromosomes}

\item{ignStrand}{}

\item{saveFlag}{a flag indicating if to save the results.}

\item{savePath}{the path where to store the results.}
}
\value{
Matrix containing read counts with regions as rows and samples as
columns.
}
\description{
count reads falling within the final regions.
}
