% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrix_functions.R
\name{countFinalRegions}
\alias{countFinalRegions}
\title{count reads falling within the final regions.}
\usage{
countFinalRegions(region_file, bam_file_path, min_carriers = 2,
  verbose = TRUE)
}
\arguments{
\item{region_file}{path to regions of interest to count reads across.}

\item{bam_file_path}{path to bam files to be used for read counting.}

\item{min_carriers}{integer indicating minumum number of replicates a region
must appear in to be utilized for count matrix construction.}
}
\value{
Matrix containing read counts with regions as rows and samples as
    columns.
}
\description{
count reads falling within the final regions.
}
