% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_call_functions.R
\name{findPeaks}
\alias{findPeaks}
\title{find peaks.}
\usage{
findPeaks(files, chr = 1:19, fraglen = 200, rlen = 100, min_win = 1,
  max_win = 100, blocksize = 10000, zthresh = 5, min_count = 0.1,
  grid = NA, filetype = "bam", save = FALSE)
}
\arguments{
\item{files}{Character vector containing paths of files to be analyzed.}

\item{chr}{Integer specifying which chromosme to limit analysis to.}

\item{fraglen}{Integer indicating the average DNA fragment length in bp}

\item{rlen}{Integer indicating the read length in bp}

\item{min_win}{Integer indicating the minimum window size to be used in constructing grid}

\item{max_win}{Integer indicating the maximum window size to be used in constructing grid}

\item{blocksize}{Integer indicating how large of a chunk of the chromosome will be analyzed at a time}

\item{zthresh}{Z-score threshold over which peaks will be retained.}

\item{min_count}{Double indicating a small amount to use as minimum count instead of zero.}

\item{grid}{Numeric vector grid}

\item{filetype}{Character, either "bam" or "bed"}

\item{save}{Boolean, save files.}
}
\value{
s.
}
\description{
find peaks.
}
\examples{
bam <- system.file("extdata", "test.bam", package = "DEScan")
peaks <- findPeaks(bam, chr = 1, filetype = "bam")
head(peaks)
}

