% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakCallFunctions.r
\name{findPeaks}
\alias{findPeaks}
\title{findPeaks}
\usage{
findPeaks(files, filetype = c("bam", "bed"), genomeName = NULL,
  binSize = 50, minWin = 50, maxWin = 1000, zthresh = 5,
  minCount = 0.1, minCompWinWidth = 100, maxCompWinWidth = 200,
  outputFolder = "Peaks", save = TRUE, force = TRUE, verbose = FALSE,
  fragmentLength = 200, onlyStdChrs = TRUE, chr = NULL)
}
\arguments{
\item{files}{Character vector containing paths of files to be analyzed.}

\item{filetype}{Character, either "bam" or "bed" indicating format of input
file.}

\item{genomeName}{the code of the genome to use as reference for the input
files. (cfr. constructBedRanges function parameters)}

\item{binSize}{Integer size in bases of the minimum window for scanning,
50 is the default.}

\item{minWin}{Integer indicating the minimum window size in bases notation.}

\item{maxWin}{Integer indicating the maximum window size in bases notation.}

\item{zthresh}{Cuttoff value for z-scores. Only windows with greater z-scores
will be kept.}

\item{minCount}{A small constant (usually no larger than one) to be added to
the counts prior to the log transformation to avoid problems with log(0).}

\item{minCompWinWidth}{minimum bases width of a comparing window for Z-score.}

\item{maxCompWinWidth}{maximum bases width of a comparing window for Z-score.}

\item{outputFolder}{A string, Name of the folder to save the Peaks (optional),
if the directory doesn't exist, it will be created. (Default is "Peaks")}

\item{save}{Boolean, if TRUE files will be saved in a "./Peaks/chr*"
directory created (if not already present) in the current working directory.}

\item{force}{a boolean flag indicating if to force output overwriting.}

\item{verbose}{if to show additional messages}

\item{fragmentLength}{Integer indicating the average DNA fragment length bp.}

\item{onlyStdChrs}{a flag to work only with standard chromosomes.
(cfr. constructBedRanges function parameters).}

\item{chr}{if not NULL, a character like "chr#" indicating the chromosomes
to use.}
}
\value{
A GRangesList where each element is a sample.
Each GRanges represents the founded peaks and attached the z-score of
the peak as mcols.
}
\description{
This function calls peaks from bed or bam inputs using a
variable window scan with a poisson model using the surrounding
maxCompWinWidth (10kb) as background.
}
\examples{
bed.files <- list.files(system.file("extdata/Bed", package = "DEScan2"),
                       full.names = T)
peaks <- findPeaks(files=bed.files[1], chr="chr19", filetype="bed",
                        fragmentLength=200,
                        binSize=50, minWin=50, maxWin=1000,
                        genomeName="mm9",
                        minCompWinWidth=5000, maxCompWinWidth=10000,
                        zthresh=5, minCount=0.1, verbose=FALSE, save=FALSE)
head(peaks)
}
