% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_call_functions.r
\name{findPeaks}
\alias{findPeaks}
\title{findPeaks}
\usage{
findPeaks(files, filetype = c("bam", "bed"), genomeName = NULL,
  binSize = 50, minWin = 1, maxWin = 20, zthresh = 5, minCount = 0.1,
  minCompWinWidth = 5000, maxCompWinWidth = 10000, outputName = "Peaks",
  save = TRUE, verbose = FALSE, fragmentLength = 200,
  onlyStdChrs = TRUE, chr = NULL)
}
\arguments{
\item{files}{Character vector containing paths of files to be analyzed.}

\item{filetype}{Character, either "bam" or "bed" indicating format of input
file.}

\item{genomeName}{the code of the genome to use as reference for the input
files. (cfr. constructBedRanges function parameters)}

\item{binSize}{Integer size in base pairs of the minimum window for scanning,
50 is the default.}

\item{minWin}{Integer indicating the minimum window size in units of 50 bp,
i.e., min_win=2 resulting in a 100 bp window.}

\item{maxWin}{Integer indicating the maximum allowed window size in units of
50 bp.}

\item{zthresh}{Cuttoff value for z-scores. Only windows with greater z-scores
will be kept.}

\item{minCount}{A small constant (usually no larger than one) to be added to
the counts prior to the log transformation to avoid
problems with log(0).}

\item{minCompWinWidth}{minimum bases width of a comparing window for Z-score}

\item{maxCompWinWidth}{maximum bases width of a comparing window for Z-score}

\item{outputName}{A string, Name of the folder to save the Peaks (optional),
if the directory doesn't exist, it will be created.
default is "Peaks"}

\item{save}{Boolean, if TRUE files will be saved in a "./Peaks/chr*"
directory created (if not already present) in the current
working directory.}

\item{fragmentLength}{Integer indicating the average DNA fragment length bp.}

\item{onlyStdChrs}{a flag to work only with standard chromosomes
(cfr. constructBedRanges function parameters)}

\item{chr}{if not NULL, a character like "chr#" indicating the
chromosomes to use}
}
\value{
A GRangesList where each element is a sample.
        each GRanges represents the founded peaks and attached the z-score
        of the peak as mcols.
}
\description{
This function calls peaks from bed or bam inputs using a
variable window scan with a poisson model using the surrounding
maxCompWinWidth (10kb) as background.
}
\examples{
TBW

}
