% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_call_functions.r
\name{get_disjoint_max_win}
\alias{get_disjoint_max_win}
\title{get_disjoint_max_win}
\usage{
get_disjoint_max_win(z0, sigwin = 20, nmax = Inf, zthresh = -Inf,
  two_sided = FALSE, verbose = FALSE)
}
\arguments{
\item{z0}{Matrix containing z scores with bins as rows and windows size as
columns}

\item{sigwin}{Integer indicating how many bins per fragment}

\item{nmax}{Integer indicating the maximum number of windows to return}

\item{zthresh}{Integer indicating the minimum z-score considered significant}

\item{two_sided}{not used argument}

\item{verbose}{verbose flag}
}
\value{
a matrix of integer containing founded peaks
}
\description{
find significant z score windows keeping the max value
               without intersections
}
\keyword{internal}
