% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_region_functions.R
\name{initMergedPeaksNames}
\alias{initMergedPeaksNames}
\title{initMergedPeaksNames}
\usage{
initMergedPeaksNames(mergedGRanges)
}
\arguments{
\item{mergedGRanges}{}
}
\value{
a granges of renamed peaks
}
\description{
given a GRanges of merged peaks assigns them new names
}
