% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_peaks.R
\name{make_grid}
\alias{make_grid}
\title{make grid for enrichment scan.}
\usage{
make_grid(bed, fraglen, rlen, min_bin = 10, min_rds = 0, grid_st = NA,
  grid_ed = NA)
}
\arguments{
\item{fraglen}{Integer indicating the average DNA fragment length in bp}

\item{rlen}{Integer indicating the read length in bp}

\item{min_bin}{Integer indicating the minimum window size to be used in
constructing grid}

\item{min_rds}{This argument is currently not used.}

\item{grid_st}{This argument is currently not used.}

\item{grid_ed}{This argument is currently not used.}

\item{path}{Path to alignement files.}
}
\value{
grid.
}
\description{
make grid for enrichment scan.
}
\keyword{internal}
