% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{readBamAsBed}
\alias{readBamAsBed}
\title{readBamAsBed: read a bam file into a bed like format.}
\usage{
readBamAsBed(file)
}
\arguments{
\item{file}{Character indicating path to bam file.}
}
\value{
GenomicRanges object.
}
\description{
readBamAsBed: read a bam file into a bed like format.
}
\keyword{internal}
