% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_call_functions.r
\name{readBamAsBed}
\alias{readBamAsBed}
\title{read a bam file into a bed like format.}
\usage{
readBamAsBed(file, chr)
}
\arguments{
\item{file}{Character indicating path to bam file.}

\item{chr}{Integer indicating which chromosome to read in.}
}
\value{
GenomicRanges object.
}
\description{
read a bam file into a bed like format.
}
\keyword{internal}
