% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{readBedFile}
\alias{readBedFile}
\title{readBedFile: read a bed file into a GenomicRanges like format.}
\usage{
readBedFile(filename)
}
\arguments{
\item{file}{Character indicating path to bam file.}
}
\value{
GenomicRanges object
}
\description{
readBedFile: read a bed file into a GenomicRanges like format.
}
\keyword{internal}
