% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{readBedFile}
\alias{readBedFile}
\title{readBedFile}
\usage{
readBedFile(filename, arePeaks = FALSE)
}
\arguments{
\item{filename}{the bed filename.}

\item{arePeaks}{a flag indicating if the the bed file represents peaks.}
}
\value{
GRanges object
}
\description{
read a bed file into a GenomicRanges like format.
             forcing UCSC format for chromosomes names.
}
\keyword{internal}
