% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{readBedFile}
\alias{readBedFile}
\title{read a bed file into a GenomicRanges like format.}
\usage{
readBedFile(filename, chr)
}
\arguments{
\item{chr}{Integer indicating which chromosome(s) to read in.}

\item{file}{Character indicating path to bam file.}
}
\value{
GenomicRanges object
}
\description{
read a bed file into a GenomicRanges like format.
}
\keyword{internal}
