% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{saveGRangesAsBed}
\alias{saveGRangesAsBed}
\title{saveGRangesAsBed}
\usage{
saveGRangesAsBed(GRanges, filepath, filename)
}
\arguments{
\item{GRanges}{the GRanges object}

\item{filepath}{the path to store the files}

\item{filename}{the name to give to the files}
}
\value{
none
}
\description{
save a GRanges object as bed file and RData file
}
\examples{
TBW
}
