% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{saveGRangesAsTsv}
\alias{saveGRangesAsTsv}
\title{saveGRangesAsTsv}
\usage{
saveGRangesAsTsv(GRanges, filepath, filename, force = FALSE,
  verbose = FALSE)
}
\arguments{
\item{GRanges}{the GRanges object}

\item{filepath}{the path to store the files}

\item{filename}{the name to give to the files}

\item{force}{force overwriting}
}
\value{
none
}
\description{
save a GRanges object as tsv file
}
\keyword{internal}
