% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setGRGenomeInfo}
\alias{setGRGenomeInfo}
\title{setGRGenomeInfo
given a genome code (i.e. "mm9","mm10","hg19","hg38") retrieve the SeqInfo of
that genome and assigns it to the input GRanges. Finally filters out those
Infos not necessary to the GRanges.}
\usage{
setGRGenomeInfo(GRanges, genomeName = NULL, verbose = FALSE)
}
\arguments{
\item{GRanges}{a GRanges object.}

\item{genomeName}{a genome code (i.e. "mm9")}

\item{verbose}{verbose output}
}
\value{
a GRanges object with the seqinfo of the genome code
}
\description{
setGRGenomeInfo
given a genome code (i.e. "mm9","mm10","hg19","hg38") retrieve the SeqInfo of
that genome and assigns it to the input GRanges. Finally filters out those
Infos not necessary to the GRanges.
}
