% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_call_functions.R
\name{window_coverage}
\alias{window_coverage}
\title{compute window coverage.}
\usage{
window_coverage(Fr, Rr, grid, fraglen = 200, rlen = 100, min_win = 1,
  max_win = 100, verbose = TRUE)
}
\arguments{
\item{Fr}{forward reads.}

\item{Rr}{reverse reads.}
}
\value{
c.
}
\description{
compute window coverage.
}
\keyword{internal}

