% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_call_functions.r
\name{window_coverage}
\alias{window_coverage}
\title{compute window coverage.}
\usage{
window_coverage(Fr, Rr, grid, fragmentLength = 200, readLength = 100,
  minWin = 1, maxWin = 100, verbose = TRUE)
}
\arguments{
\item{Fr}{forward reads.}

\item{Rr}{reverse reads.}
}
\value{
cmat.
}
\description{
compute window coverage.
}
\keyword{internal}
